requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
# dùng khoảng phiên bản đã có trên PyPI
yt-dlp>=2024.10.0,<2026.0.0
pydantic==2.9.2
tenacity==9.0.0
